libct/cg/sd: fix dbus error handling
This fixes isDbusError function, introduced by commit 6e8f491. Due to a
type error it was not working at all.

This also fixes the whole "retry on dbus disconnect" logic.

This also fixes a regression in startUnit (and cgroupManager.Apply()),
which should never return "unit already exists" error but it did.

Fixes: 6e8f491
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>