systemd: Write rounded CPU quota to cgroupfs
When CPU quota is updated, the value is converted to CPUQuotaPerSecUSec
property for passing to systemd. The value can be rounded in the
following cases:
- The value is rounded up to the nearest 10ms.
- Depending on CPU period, the value may be rounded during division.

Because of this rounding, systemd and systemd driver may write
different values to cgroupfs. In order to avoid this inconsistency,
this fix makes systemd driver write the same rounded value to cgroupfs.

Even if systemd writes CPU quota and CPU period to cgroupfs, systemd
driver still needs to write to cgroupfs for a case where CPU period
is updated to less than the minimum value. In this case, systemd accepts
this value by adjusting CPU period while direct update by systemd driver
fails. In order to keep this case invalid, systemd driver still needs to
update cgroupfs.

Signed-off-by: Hironori Shiina <shiina.hironori@gmail.com>