Add .github/dependabot.yml
Make sure our github actions stay up to date.

Since this is a package (set of packages), we do not to bump
dependencies to the latest versions automatically, but only
when it is really required, thus sticking to Go's minimum version
selection rules (essentially so that users of these packages
are not required to unnecessarily update their dependencies).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>