fs2: add iocost statistics
When the iocost controller is enabled it will emit statistics about its
usage in io.stat. Export this when available. This is a no-op when
iocost is not enabled.

There is a comment that we only expose data which directly maps to
cgroups v1. We're already breaking this by adding PSI metrics, so reword
this to just indicate that we currently don't recognize the other
fields.

Signed-off-by: Morten Hein Tiljeset <morten@tiljeset.com>