fs: improve cpuacct.usage_all parsing
First, we can reasonably expect the first three fields of
cpuacct.usage_all to be "cpu user system", even for future kernels,
and if we see something different, it doesn't make sense to continue.
So check that the header is as expected, and error out otherwise.

Second, if we have more than 3 values, and we've checked that the first
3 are as expected ("cpu user system"), we can safely ignore extra
columns. This fixes an issue on a custom kernel from Tencent, which
adds a few extra columns (see [1]), as reported in issue 46.

Add tests for both cases.

Fixes issue 46.

[1]: https://github.com/OpenCloudOS/TencentOS-kernel-0/commit/0b667819c3aaa9c8ac904c6b03256b86be93fc05
Reported-by: vimiix <i@vimiix.com>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>